+++ .github/actions/describe/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "hotfix",
+  "name": "describe",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
