--- .github/actions/hotfix/action.yml
@@ -1,16 +1,0 @@
-name: "Hotfix"
-description: "This is an emergency action for quick fixes to released code."
-inputs:
-  token:
-    description: "GitHub token to authenticate API requests."
-    required: true
-outputs:
-  message:
-    description: "A commit message describing the hotfix."
-  version:
-    description: "The new version number after applying the hotfix."
-  comment:
-    description: "A comment notifying the author of the hotfix status."
-runs:
-  using: "node16"
-  main: "dist/index.js"
